requirements.txt
aioredis==1.3.1
apparmor==3.0.4
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
btrfsutil==5.16.2
CacheControl==0.12.6
cachetools==5.0.0
caffeine-ng==3.5.1.post71+g77611ac
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.4
colorama==0.4.4
configobj==5.1.0.dev0
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==36.0.1
daphne==3.0.2
distlib==0.3.4
distro==1.7.0
Django==4.0.3
djangorestframework==3.13.1
dnspython==2.2.0
docopt==0.6.2
duplicity==0.8.21
env==0.1.0
ewmh==0.1.6
fasteners==0.16.3
filelock==3.4.2
future==0.18.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gufw==21.4.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.20.3
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.1
incremental==21.3.0
layoutswitcherlib==0.8.24
LibAppArmor==3.0.4
lit==13.0.1.dev0
louis==3.21.0
manjaro-sdk==0.1
Markdown==3.3.6
more-itertools==8.10.0
msgpack==1.0.3
netsnmp-python==1.0a1
npyscreen==4.10.5
oauth2client==4.1.3
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
paramiko==2.9.2
pep517==0.12.0
Pillow==9.0.1
ply==3.11
progress==1.6
protobuf==3.19.4
pulsectl==22.1.3
pwquality==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
PyDrive==1.3.1
PyGObject==3.42.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PySide6==6.2.3
pysmbc==1.0.23
python-xlib==0.31
pytz==2021.3
pyxdg==0.27
PyYAML==6.0
ranger-fm==1.9.3
reportlab==3.6.6
requests==2.27.1
requests-file==1.5.1
resolvelib==0.5.5
retrying==1.3.3
rsa==4.8
service-identity==21.1.0
setproctitle==1.2.2
shiboken6==6.2.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
systemd-python==234
team==1.0
tldextract==3.2.0
toml==0.10.2
tomli==2.0.0
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
ueberzug==18.1.9
ufw==0.36.1
uritemplate==3.0.1
urllib3==1.26.8
validate==5.1.0.dev0
webencodings==0.5.1
zipp==3.7.0
zope.interface==5.4.0
